feat: complete Phase 3b — MSSQL, Oracle, SQLite connectors with 20 cross-database migration scripts
Phase 3b adds three new connectors (MSSQL, Oracle, SQLite) and validates
all 20 bidirectional migration pairs across the five supported databases.

Connectors:
- MSSQL connector (pymssql): source + sink with binary column handling
  via CONVERT(), NVARCHAR(MAX) index narrowing, system table filtering
- Oracle connector (oracledb 3.x): source + sink with IDENTITY detection,
  NUMBER(p,0) → integer mapping, system table filtering, FK ON DELETE
- SQLite connector: source + sink with full type mapping

Cross-cutting fixes applied during integration testing:
- MySQL data reader: read column flags from cursor._result.fields (not
  cursor.description) for unsigned type detection
- MySQL data reader: use charsetnr to disambiguate TEXT vs BINARY columns
  sharing the same type codes on the wire
- MSSQL/Oracle data readers: query INFORMATION_SCHEMA for real type names
  instead of relying on cursor.description type objects
- MSSQL schema reader: build full type strings (e.g. nvarchar(20)) from
  CHARACTER_MAXIMUM_LENGTH, strip wrapping parens from defaults
- Oracle schema reader: filter system tables via user_tables NOT LIKE '%$%',
  detect ISEQ$$ identity sequences
- All sink connectors: recover varchar length from source data_type when
  Arrow string type loses it; quote bare string defaults
- PG connector: use bigserial for int64 auto-increment columns
- Shared value_coercion module and default_translation module
- Docker: Oracle image switched to 23-slim-faststart, dedicated bani_test
  user instead of SYSTEM schema
- Migration scripts renumbered 01–20 in logical order with run_all_migrations.sh

All 20 migration scripts pass: MySQL, PostgreSQL, MSSQL, Oracle, SQLite
in every direction.

+++ src/bani/connectors/mssql/type_mapper.py
@@ -136,7 +136,7 @@ class MSSQLTypeMapper:
             # Boolean
             "bool": "BIT",
             # Integer types
-            "int8": "TINYINT",
+            "int8": "SMALLINT",  # MSSQL TINYINT is unsigned (0-255)
             "int16": "SMALLINT",
             "int32": "INT",
             "int64": "BIGINT",
